# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_qun100_image": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "auth_token": {
            "maxLength": 4096,
            "minLength": 60,
            "title": "Auth Token",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "required": [
          "auth_token",
          "file"
        ],
        "title": "Body_upload_qun100_image",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QunCatalogItem": {
        "properties": {
          "cid": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Cid",
            "type": "string"
          },
          "type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "cid",
          "type",
          "value"
        ],
        "title": "QunCatalogItem",
        "type": "object"
      },
      "QunFormSummary": {
        "properties": {
          "form_id": {
            "title": "Form Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "form_id",
          "title",
          "status",
          "version"
        ],
        "title": "QunFormSummary",
        "type": "object"
      },
      "QunImageResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "content_type",
          "size"
        ],
        "title": "QunImageResponse",
        "type": "object"
      },
      "QunImmediateSubmitRequest": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "auth_token": {
            "format": "password",
            "title": "Auth Token",
            "type": "string",
            "writeOnly": true
          },
          "catalogs": {
            "items": {
              "$ref": "#/components/schemas/QunCatalogItem"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Catalogs",
            "type": "array"
          },
          "confirmation_version": {
            "title": "Confirmation Version",
            "type": "string"
          },
          "form_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Form Version"
          },
          "title": {
            "default": "",
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "required": [
          "auth_token",
          "form_version",
          "catalogs",
          "confirmation_version"
        ],
        "title": "QunImmediateSubmitRequest",
        "type": "object"
      },
      "QunPreviewRequest": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "auth_token": {
            "format": "password",
            "title": "Auth Token",
            "type": "string",
            "writeOnly": true
          },
          "custom_fields": {
            "additionalProperties": true,
            "title": "Custom Fields",
            "type": "object"
          },
          "defaults": {
            "$ref": "#/components/schemas/QunUserDefaults"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "required": [
          "auth_token"
        ],
        "title": "QunPreviewRequest",
        "type": "object"
      },
      "QunPreviewResponse": {
        "properties": {
          "catalogs": {
            "items": {
              "$ref": "#/components/schemas/QunCatalogItem"
            },
            "title": "Catalogs",
            "type": "array"
          },
          "form_id": {
            "title": "Form Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "form_id",
          "title",
          "version",
          "catalogs"
        ],
        "title": "QunPreviewResponse",
        "type": "object"
      },
      "QunResolveRequest": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "auth_token": {
            "format": "password",
            "title": "Auth Token",
            "type": "string",
            "writeOnly": true
          },
          "input": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "required": [
          "auth_token",
          "input"
        ],
        "title": "QunResolveRequest",
        "type": "object"
      },
      "QunSubmitResponse": {
        "properties": {
          "form_id": {
            "title": "Form Id",
            "type": "string"
          },
          "submitted": {
            "title": "Submitted",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "form_id",
          "title",
          "submitted"
        ],
        "title": "QunSubmitResponse",
        "type": "object"
      },
      "QunUserDefaults": {
        "properties": {
          "default_address": {
            "default": "",
            "maxLength": 300,
            "title": "Default Address",
            "type": "string"
          },
          "default_lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Lat"
          },
          "default_lng": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Lng"
          },
          "display_name": {
            "default": "",
            "maxLength": 100,
            "title": "Display Name",
            "type": "string"
          }
        },
        "title": "QunUserDefaults",
        "type": "object"
      },
...
      "TokenRequest": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "auth_token": {
            "format": "password",
            "title": "Auth Token",
            "type": "string",
            "writeOnly": true
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          }
        },
        "required": [
          "auth_token"
        ],
        "title": "TokenRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Nanyee API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/qun/forms": {
      "post": {
        "operationId": "list_qun100_forms",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QunFormSummary"
                  },
                  "title": "Response List Qun100 Forms",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Forms",
        "tags": [
          "qun100"
        ]
      }
    },
    "/api/v1/qun/forms/resolve": {
      "post": {
        "operationId": "resolve_qun100_form",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QunResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QunFormSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Form",
        "tags": [
          "qun100"
        ]
      }
    },
    "/api/v1/qun/forms/{form_id}/preview": {
      "post": {
        "operationId": "preview_qun100_form",
        "parameters": [
          {
            "in": "path",
            "name": "form_id",
            "required": true,
            "schema": {
              "title": "Form Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QunPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QunPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Form",
        "tags": [
          "qun100"
        ]
      }
    },
    "/api/v1/qun/forms/{form_id}/submit": {
      "post": {
        "operationId": "submit_qun100_form_immediately",
        "parameters": [
          {
            "in": "path",
            "name": "form_id",
            "required": true,
            "schema": {
              "title": "Form Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QunImmediateSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QunSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Form Immediately",
        "tags": [
          "qun100"
        ]
      }
    },
    "/api/v1/qun/images": {
      "post": {
        "operationId": "upload_qun100_image",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_qun100_image"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QunImageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Image",
        "tags": [
          "qun100"
        ]
      }
    },
    "/api/v1/qun/token/verify": {
      "post": {
        "operationId": "verify_qun100_token",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Verify Qun100 Token",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Token",
        "tags": [
          "qun100"
        ]
      }
    },
...
  }
}